warnings fixed in brandisher and simdb
tbl array stats are no longer created dynamically

+++ simdb.hpp
@@ -1596,7 +1596,7 @@ public:
     }
 
     //auto cncrHashSize = CncrHsh::sizeBytes(blockCount);
-    uint64_t cncrHashSize = CncrHsh::sizeBytes(s_blockCount->load());
+    uint64_t cncrHashSize = CncrHsh::sizeBytes((u32)s_blockCount->load());
     new (&s_cs) CncrStr( ((u8*)m_mem.data())+cncrHashSize+OffsetBytes(), 
                                  (u32)s_blockSize->load(), 
                                  (u32)s_blockCount->load(),
